requirements.txt
python-dotenv
pandas
PyMuPDF
chromadb
sentence-transformers
transformers
streamlit
tqdm
rich
requests
fitz
langchain